Changed parse_single_token() to warn on invalid entries instead of panic.
